Implement sadfilesys::write()

+++ include/sadfs/sadfsd/sadfilesys.hpp
@@ -48,12 +48,12 @@ private:
     int open(char const* path, fuse_file_info* fi);
 
     // read data from an open file
-    int read(char const* path, char* buf, size_t size, off_t offset,
+    int read(char const* path, char* buf, uint32_t size, uint32_t offset,
              fuse_file_info* fi);
 
     // write to an open file
-    int write(char const* path, const char* buf, size_t size, off_t offset,
-    	      fuse_file_info* fi);
+    int write(char const* path, const char* buf, uint32_t size,
+    	      uint32_t offset, fuse_file_info* fi);
 
     fuse_operations operations_{};
